LayoutTests/fast/loader/refresh-iframe-base-blank.html

<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    testRunner.dumpFrameLoadCallbacks();
}

</script>

<body>
<iframe src="resources/refresh-iframe-base-blank-frame.html"></iframe>
</body>
